requirements.txt
asgiref==3.4.1
Django==3.2.5
django-countries==7.2.1
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
django-tables2==2.4.0
phonenumbers==8.12.27
pytz==2021.1
sqlparse==0.4.1
